scripts/requirements.txt
# 信息源蒸馏脚本依赖（conda 环境 ikb_env）
openai>=1.0.0
python-dotenv>=1.0.0
feedparser>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
